requirements.txt
# FastAPI and Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-dotenv==1.0.0

# Pydantic
pydantic==2.5.3
pydantic-settings==2.1.0

# Database - DOWNGRADE supabase to fix httpx compatibility
supabase==2.0.3
pymongo==4.6.1

# File processing
PyPDF2==3.0.1
python-docx==1.1.0

# HTTP client - Match with supabase 2.0.3
httpx==0.24.1

# Additional dependencies
typing-extensions==4.9.0
annotated-types==0.6.0